refactor: Moved learner model
