+++ src/app/modules/users/learner.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Learner } from './learner.model';
+import { Learner } from './model/learner.model';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../environments/environment';
 import { tap } from 'rxjs/operators';
